Move Prism helpers to prism module
No reason to keep these here, imo.
