+++ client/helpers/index.ts
@@ -1,3 +1,2 @@
 export { default as i18n } from './i18n';
 export { default as link } from './link';
-export { default as prismSlug } from './prismSlug';
